+++ stac/__init__.py
@@ -14,4 +14,4 @@ stac
 Smarter Travel Artifactory Client.
 """
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'